Fix crash when going back to lobby after a game
The crash happened because of the ShockwaveManager.Instance.DisconnectSuit() call because a life gain pattern was ran at the same time.
We now initialize the Shockwave suit when the mod starts (instead of when entering FP view, which was sometimes missing early elevator haptics), and we now only disconnect the suit when quitting the game.